Crate owned_borrowed_trait

An abstraction over an owned value or borrowed reference based on traits and marker types instead of enums.

Structs

Borrowed

A wrapper that indicates the contained value is a borrowed reference to a value of type T.

Owned

A wrapper that indicates the contained value is an owned value of type T.

Traits

OwnedOrBorrowed

A trait for abstracting over an owned value or borrowed reference to a type T.