Trait owned_borrowed_trait::OwnedOrBorrowed

pub trait OwnedOrBorrowed<T>: Sealed {
    type Inner: Borrow<T>;

    const OWNED: bool;

    // Required methods
    fn into_inner(self) -> Self::Inner;
    fn as_inner(&self) -> &Self::Inner;
}

A trait for abstracting over an owned value or borrowed reference to a type T.

You cannot implement this trait; it can only be used with Owned or Borrowed.

The Owned and Borrowed wrapper types implement the following traits:

AsRef.

Deref where Target = T.

Required Associated Types

type Inner: Borrow<T>

The inner type of the owned value or borrowed reference.

Required Associated Constants

const OWNED: bool

true if the wrapper type contains an owned value, i.e., for Owned.

false if the wrapper type contains a borrowed reference, i.e., for Borrowed.

Required Methods

fn into_inner(self) -> Self::Inner

Consumes this wrapper type and returns the contained value or borrowed reference.

fn as_inner(&self) -> &Self::Inner

Returns a reference to the inner value.